clean interface.h
