--- include/fti-int/dcp-func.h
@@ -1,1 +1,0 @@
-#ifndef DCP_POSIX_FUNC
