--- include/fti-int/func.h
@@ -1,10 +1,0 @@
-#ifndef FUNC_H
-#define FUNC_H
-
-#include <fti-int/ftiff-func.h>
-#include <fti-int/stage-func.h>
-#include <fti-int/icp-func.h>
-#include <fti-int/hdf5-func.h>
-#include <fti-int/utility-func.h>
-
-#endif // FUNC_H
