+++ include/fti-int/hdf5-fti.h
@@ -1,7 +1,8 @@
 #ifndef HDF5_FUNC_H
 #define HDF5_FUNC_H
 
-#ifdef ENABLE_HDF5
+#include <fti-int/types.h>
+
 int FTI_WriteHDF5(FTIT_configuration* FTI_Conf, FTIT_execution* FTI_Exec,
                   FTIT_topology* FTI_Topo, FTIT_checkpoint* FTI_Ckpt,
                   FTIT_dataset* FTI_Data);
@@ -26,6 +27,5 @@ void FTI_OpenGroup(FTIT_H5Group* ftiGroup, hid_t parentGroup, FTIT_H5Group** FTI
 void FTI_CloseGroup(FTIT_H5Group* ftiGroup, FTIT_H5Group** FTI_Group);
 int FTI_CreateGlobalDatasets( FTIT_execution* FTI_Exec );
 int FTI_CloseGlobalDatasets( FTIT_execution* FTI_Exec );
-#endif
 
 #endif // HDF5_FUNC_H
