--- src/utility.h
@@ -1,47 +1,0 @@
-#ifndef __UTILITY__
-#define __UTILITY__
-#include <string.h>
-
-#include "IO/ftiff.h"
-
-
-
-typedef struct
-{
-	FTIT_configuration* FTI_Conf;
-	FTIT_topology *FTI_Topo;
-	MPI_Offset offset;
-	int err;
-	MPI_Info info;
-	MPI_File pfh;
-	char flag;
-} WriteMPIInfo_t;
-
-typedef struct{
-	FILE *f;
-	size_t offset;
-	char flag;
-}WritePosixInfo_t;
-
-
-// Wrappers around MPIO
-int FTI_MPIOOpen(char *fn, void *fileDesc);
-int FTI_MPIOClose(void *fileDesc);
-int FTI_MPIOWrite(void *src, size_t size, void *fileDesc);
-int FTI_MPIORead(void *src, size_t size, void *fileDesc);
-
-//Wrappers around POSIX IO
-int FTI_PosixOpen(char *fn, void *fileDesc);
-int FTI_PosixClose(void *fileDesc);
-int FTI_PosixWrite(void *src, size_t size, void *fileDesc);
-int FTI_PosixRead(void *src, size_t size, void *fileDesc);
-int FTI_PosixSync(void *fileDesc);
-int FTI_PosixSeek(size_t pos, void *fileDesc);
-
-
-int copyDataFromDevive(FTIT_execution* FTI_Exec, FTIT_dataset* FTI_Data);
-
-#ifdef ENABLE_SIONLIB 
-int write_sion(void *src, size_t size, void *opaque);
-#endif
-#endif
